backend/requirements.txt
autopep8
Django==4.1.*
djangorestframework
djangorestframework-simplejwt
djangorestframework-camel-case
django-rest-polymorphic
django-cors-headers
django-filter
django-url-filter
django-guardian
markdown
drf-yasg
python-dotenv